refactor: Move TokoPicker component under Home page
