+++ src/pages/Home/TokoPicker/index.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 
-import FilterInput from'../FilterInput';
-import FilterCards from '../FilterCards';
-import { tokos } from '../../models';
+import FilterInput from'../../../components/FilterInput';
+import FilterCards from '../../../components/FilterCards';
+import { tokos } from '../../../models';
 import './TokoPicker.css';
 
 export default class TokoPicker extends Component {
